fix register bugs

+++ src/pages/Register.js
@@ -5,6 +5,7 @@ import { loginThunk, registerThunk } from "../services/users/users-thunks";
 import { initFollowThunk } from "../services/thunks/follow-thunk";
 import { createPlaylist } from "../services/playlist-service";
 import "./Register_styles.css";
+import { ADMINCODE } from "../utils/URL";
 
 const Register = () => {
   const dispatch = useDispatch();
@@ -14,7 +15,14 @@ const Register = () => {
   const [email, setEmail] = useState("");
   const [gender, setGender] = useState("");
   const [cellphone, setCellphone] = useState("");
-    const [isAdmin, setIsAdmin] = useState(false);
+  const [isAdmin, setIsAdmin] = useState(false);
+  const [adminCode, setAdminCode] = useState("");
+  const [adminError, setAdminError] = useState(false);
+  const [userNameAlert, setUserNameAlert] = useState(false);
+  const [passwordAlert, setPasswordAlert] = useState(false);
+  const [phoneAlert, setPhoneAlert] = useState(false);
+  const [emailAlert, setEmailAlert] = useState(false);
+  const [genderAlert, setGenderAlert] = useState(false);
 
   const addPlaylist = async (userId) => {
     const newPlaylist = {
@@ -28,25 +36,70 @@ const Register = () => {
     await createPlaylist({ playlist: newPlaylist, cnt: 1 });
   };
 
+  const checkInfo = () => {
+    setUserNameAlert(false);
+    setPasswordAlert(false);
+    setPhoneAlert(false);
+    setEmailAlert(false);
+    setGenderAlert(false);
+    if (userName === "") {
+      setUserNameAlert(true);
+      return false;
+    }
+    if (email === "") {
+      setEmailAlert(true);
+      return false;
+    }
+    if (cellphone === "") {
+      setPhoneAlert(true);
+      return false;
+    }
+    if (password === "") {
+      setPasswordAlert(true);
+      return false;
+    }
+    if (gender === "") {
+      setGenderAlert(true);
+      return false;
+    }
+    return true;
+  };
+
   const register = async () => {
+    setAdminError(false);
     try {
+      if (isAdmin && adminCode !== ADMINCODE) {
+        setAdminError(true);
+        return;
+      }
+      setAdminCode("");
+      // check each field before submit
+      const isValid = checkInfo();
+      if (!isValid) {
+        return;
+      }
       localStorage.clear();
       await dispatch(
-        registerThunk({ userName, password, email, cellphone, gender, isAdmin })
-      );
+        registerThunk({
+          userName,
+          password,
+          email,
+          cellphone,
+          gender,
+          isAdmin,
+        })
+      ).then((res) => {
+        const user_id = res.payload._id;
+        // add one default playlist for new user
+        addPlaylist(user_id);
+        // add empty followeeList for new user
+        dispatch(initFollowThunk(user_id));
+      });
       // navigate("/login");
       await dispatch(loginThunk({ userName, password })).then((res) => {
-        const user_id = JSON.parse(
-          window.localStorage.getItem("currentUser")
-        )._id;
+        const user_id = res.payload._id;
         navigate(`/home?_id=${user_id}`);
       });
-      const user_id = JSON.parse(
-        window.localStorage.getItem("currentUser")
-      )._id;
-      dispatch(initFollowThunk(user_id));
-      // add one default playlist for user
-      addPlaylist(user_id);
     } catch (error) {
       alert("User name or email already exists!");
     }
@@ -65,12 +118,20 @@ const Register = () => {
               <h1 className={`text-white fw-bold`}>Create new account</h1>
 
               <div className={``}>
-                <label
-                  htmlFor="userName"
-                  className="mt-2 text-warning fw-bold mb-2"
-                >
-                  UserName
-                </label>
+                <div className={`row w-100 p-0 m-0 d-flex align-items-center`}>
+                  <label
+                    htmlFor="userName"
+                    className="mt-2 text-warning fw-bold mb-2 col-4 p-0"
+                  >
+                    UserName
+                  </label>
+                  {userNameAlert && (
+                    <p className={`mb-0 text-danger col p-0`}>
+                      Please enter username
+                    </p>
+                  )}
+                </div>
+
                 <input
                   id="userName"
                   placeholder="Enter a userName*"
@@ -85,12 +146,20 @@ const Register = () => {
               </div>
 
               <div className={`mt-3`}>
-                <label
-                  htmlFor="email"
-                  className="mt-2 text-warning fw-bold mb-2"
-                >
-                  Email
-                </label>
+                <div className={`row w-100 p-0 m-0 d-flex align-items-center`}>
+                  <label
+                    htmlFor="email"
+                    className="mt-2 text-warning fw-bold mb-2 col-4 p-0"
+                  >
+                    Email
+                  </label>
+                  {emailAlert && (
+                    <p className={`mb-0 text-danger col p-0`}>
+                      Please enter email
+                    </p>
+                  )}
+                </div>
+
                 <input
                   id="email"
                   placeholder="Enter your email*"
@@ -105,12 +174,20 @@ const Register = () => {
               </div>
 
               <div className={`mt-3`}>
-                <label
-                  htmlFor="phone"
-                  className="mt-2 text-warning fw-bold mb-2"
-                >
-                  Phone
-                </label>
+                <div className={`row w-100 p-0 m-0 d-flex align-items-center`}>
+                  <label
+                    htmlFor="phone"
+                    className="mt-2 text-warning fw-bold mb-2 col-4 p-0"
+                  >
+                    Phone
+                  </label>
+                  {phoneAlert && (
+                    <p className={`mb-0 text-danger col p-0`}>
+                      Please enter phone number
+                    </p>
+                  )}
+                </div>
+
                 <input
                   id="phone"
                   placeholder="Enter your cellphone*"
@@ -125,12 +202,20 @@ const Register = () => {
               </div>
 
               <div className={`mt-3`}>
-                <label
-                  htmlFor="password"
-                  className="mt-2 text-warning fw-bold mb-2"
-                >
-                  Password
-                </label>
+                <div className={`row w-100 p-0 m-0 d-flex align-items-center`}>
+                  <label
+                    htmlFor="password"
+                    className="mt-2 text-warning fw-bold mb-2 col-4 p-0"
+                  >
+                    Password
+                  </label>
+                  {passwordAlert && (
+                    <p className={`mb-0 text-danger col p-0`}>
+                      Please enter password
+                    </p>
+                  )}
+                </div>
+
                 <input
                   id="password"
                   placeholder="Create a password*"
@@ -145,7 +230,17 @@ const Register = () => {
               </div>
 
               <div className={`mt-3`}>
-                <label className="mt-2 text-warning fw-bold mb-2">Gender</label>
+                <div className={`row w-100 p-0 m-0 d-flex align-items-center`}>
+                  <label className="mt-2 text-warning fw-bold mb-2 col-4 p-0">
+                    Gender
+                  </label>
+                  {genderAlert && (
+                    <p className={`mb-0 text-danger col p-0`}>
+                      Please select gender
+                    </p>
+                  )}
+                </div>
+
                 <div className={`mt-2 d-flex justify-content-start`}>
                   <label className={`text-muted fw-bold`}>
                     <input
@@ -184,31 +279,55 @@ const Register = () => {
               </div>
 
               <div className={`mt-3`}>
-                <label className="mt-2 text-warning fw-bold mb-2">User Type</label>
-                <div className={`mt-2 d-flex justify-content-start`}>
+                <label className="mt-2 text-warning fw-bold mb-2 col-4 p-0">
+                  User Type
+                </label>
+
+                <div
+                  className={`mt-2 d-flex justify-content-start align-items-center`}
+                >
                   <label className={`text-muted fw-bold`}>
                     <input
-                        type="radio"
-                        name="userType"
-                        value="false"
-                        onChange={(e) => {
-                          setIsAdmin(e.target.value === "true");
-                        }}
+                      type="radio"
+                      name="userType"
+                      value="false"
+                      onChange={(e) => {
+                        setIsAdmin(e.target.value === "true");
+                        setAdminError(false);
+                        setAdminCode("");
+                      }}
                     />
                     Regular
                   </label>
-                  <label className={`ms-3 text-muted fw-bold`}>
+                  <label className={`ms-2 text-muted fw-bold`}>
                     <input
-                        type="radio"
-                        name="userType"
-                        value="true"
-                        onChange={(e) => {
-                          setIsAdmin(e.target.value === "true");
-                        }}
+                      type="radio"
+                      name="userType"
+                      value="true"
+                      onChange={(e) => {
+                        setIsAdmin(e.target.value === "true");
+                      }}
                     />
                     Admin
                   </label>
+                  {isAdmin && (
+                    <label>
+                      <input
+                        type="text"
+                        name="adminCode"
+                        value={adminCode}
+                        placeholder="Enter Code"
+                        onChange={(e) => {
+                          setAdminCode(e.target.value);
+                        }}
+                        className={`form-control admin-control-input fw-bold text-white ms-1`}
+                      />
+                    </label>
+                  )}
                 </div>
+                {adminError && (
+                  <p className={`text-danger mb-0`}>Wrong Admin Code</p>
+                )}
               </div>
 
               <div className={`mt-5 d-flex justify-content-center `}>
